refactor(run-e2e): rename commands to lib
