+++ automation/run-e2e/lib/setup-test-project.mjs
