+++ automation/run-e2e/lib/update-widget.mjs
